separate Character Manifest Data from character manager class
